+++ app/controllers/transports/transportsController.js
@@ -1,6 +1,7 @@
 const basicauth = require('basic-auth')
 const responses = require('../../config/responses/responses')
 const Transport = require('../../models/TrackedRoutes')
+const Tracker = require('../../models/Trackers')
 const trackerController = require('../../controllers/tracker/trackerController')
 
 module.exports = {
@@ -23,15 +24,53 @@ module.exports = {
         let newTransport = new Transport(req.body)
         newTransport.save()
         .then(transport => {
-            if (req.body.trackerId !== undefined || req.body.trackerId !== '') {
-                trackerController.linkVehicleToTracker(req.body.trackerId, transport._id)
-                transport.tracker = req.body.trackerId
-                transport.save()
+
+            if (req.body.trackerSerial) {
+                console.log('preenchido')
+                trackerController.linkVehicleToTracker(req.body.trackerSerial, transport._id)
+
+                Tracker.findOne({serialKey : req.body.trackerSerial})
+                .then(tracker => {
+                    transport.tracker = tracker._id
+                    transport.save()
+                    .then(result => {
+                        return res.status(200).json({
+                            success: true, message: "transporte e tracker cadastrado com sucesso", data: result
+                        })
+                    })
+                })
+            } else {
+                return res.status(200).json({
+                    success: true, message: "transporte cadastrado com sucesso", data: transport
+                })
             }
+
+
+
+            // Verifica se o transport possui o campo tracker preenchido - tracker ok
+            // if(!transport.tracker == undefined || !transport.tracker == ''){
+            //     return res.status(200).json({
+            //         success: true, message: "transporte cadastrado com sucesso.", data: transport
+            //     })
+            // }
+
+            // //Verifica se o campo trackerSerial do formulario esta preenchido - 
+            // if (!req.body.trackerSerial == undefined || !req.body.trackerSerial == '') {
+            //     // console.log('aqui')
+            //     trackerController.linkVehicleToTracker(req.body.trackerSerial, transport._id)
+                
+            //     // Tracker.find({trackerSerial: trackerSerial})
+            //     Tracker.find({serialKey : req.body.trackerSerial})
+            //     .then(tracker => {
+            //         console.log(tracker)
+            //         transport.tracker = tracker._id
+            //         transport.save()
+            //     })
+            // }
             
-            return res.status(200).json({
-                success: true, message: "transporte cadastrado com sucesso", data: transport
-            })
+            // return res.status(200).json({
+            //     success: true, message: "transporte cadastrado com sucesso,x", data: transport
+            // })
 
         })
         .catch(err => {
@@ -47,35 +86,46 @@ module.exports = {
         
         let transportId = req.params.id
 
-        if (req.body.trackerId == undefined || req.body.trackerId == '') {
+        if (req.body.trackerSerial == undefined || req.body.trackerSerial == '') {
             return res.status(responses.BAD_REQUEST).json({
                 code: responses.BAD_REQUEST,
-                error: "invalid_body_register",
+                error: "invalid_body_parameter",
                 error_description: "tracker is required"
             })
         }
 
-        Transport.findById(transportId)
-        .then(transport => {
-            trackerController.linkVehicleToTracker(req.body.trackerId, transportId)
-            transport.tracker = req.body.trackerId
-            transport.save()
-            .then(result => {
-                return res.status(200).json({
-                    success: true, message: "tracker vinculado com sucesso", data: result
+        Tracker.findOne({serialKey : req.body.trackerSerial})
+        .then(tracker => {
+            if (!tracker.vehicle == null || !tracker.vehicle == '') {
+                return res.status(400).json({
+                    code: 400, error: "invalid_insert", error_description: "Tracker ja vinculado."
+                })
+            }
+            //Filtra o transport e vincula o tracker em tranports e trackers
+            Transport.findById(transportId)
+            .then(transport => {
+                // console.log(transport)
+                trackerController.linkVehicleToTracker(req.body.trackerSerial, transportId)
+                transport.tracker = tracker._id
+                transport.save()
+                .then(result => {
+                    return res.status(200).json({
+                        success: true, message: "tracker vinculado com sucesso", data: result
+                    })
+                })
+                .catch(err => {
+                    return res.status(400).json({
+                        code: 400, error: "invalid_insert", error_description: "erro ao inserir o dado na base"
+                    })
                 })
             })
             .catch(err => {
                 return res.status(400).json({
-                    code: 400, error: "invalid_insert", error_description: "erro ao inserir o dado na base"
+                    code: 400, error: "invalid_insert", error_description: "erro ao carregar dados da base"
                 })
             })
         })
-        .catch(err => {
-            return res.status(400).json({
-                code: 400, error: "invalid_insert", error_description: "erro ao carregar dados da base"
-            })
-        })
+        .catch(err => console.log(err))
     },
 
     getAllTransports: (req, res, next) => {