+++ app/models/TrackedRoutes.js
@@ -6,6 +6,7 @@ const Routes = new Schema({
     type: {type: String, required: true},
     capacity: {type: String, required: true},
     vehiclePlate: { type: String, required: true },
+    trackerSerial: {type: String},
     tracker: {type: Schema.Types.ObjectId, ref: 'Trackersv1'},
     coordinates: {type: Array},
     thirdCompany: {type: String},