+++ app/models/Trackers.js
@@ -4,10 +4,11 @@ const Schema = mongoose.Schema;
 const Tracker = new Schema({
 
     serialKey: {type: String, required: true},
+    trackerModel: { type: String},
     notes: { type: String},
     vehicle: {type: Schema.Types.ObjectId, ref: 'Routesv1'},
     metadata: {type: Object},
-    coordinates: {type: Array}
+    status: {type: Boolean}
 
 }, {usePushEach: true});